requirements.txt
asgiref==3.2.7
astroid==2.4.0
attrs==20.2.0
autopep8==1.4.4
base58==2.0.1
beautifulsoup4==4.9.1
bitarray==1.2.2
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
cytoolz==0.10.1
dateutils==0.6.8
dj-database-url==0.5.0
Django==3.0.3
django-bootstrap3==12.1.0
django-bootstrap4==1.1.1
django-clearcache==1.1.1
django-crispy-forms==1.9.1
django-heroku==0.3.1
django-rest-framework-braces==0.3.4
djangorestframework==3.11.0
eth-abi==2.1.1
eth-account==0.5.3
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.0
eth-typing==2.2.2
eth-utils==1.9.5
gunicorn==20.0.4
hexbytes==0.2.1
idna==2.10
importlib-metadata==1.7.0
ipfshttpclient==0.6.1
isort==4.3.21
jsonschema==3.2.0
lazy-object-proxy==1.4.3
lru-dict==1.1.6
mccabe==0.6.1
multiaddr==0.0.9
netaddr==0.8.0
parsimonious==0.8.1
Pillow==7.1.2
protobuf==3.13.0
psycopg2==2.8.6
pycodestyle==2.5.0
pycryptodome==3.9.8
pylint==2.5.0
pypiwin32==223
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.1
pywin32==228
requests==2.24.0
rlp==2.0.0a1
rusty-rlp==0.1.15
six==1.14.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.0
toolz==0.10.0
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
varint==1.0.2
waitress==1.4.4
web3==5.12.1
websockets==8.1
whitenoise==5.2.0
wincertstore==0.2
wrapt==1.11.2
zipp==3.1.0
